refactor(installer): move dylint tool installation helpers to install_flow
Extract dylint tools installation functions from main.rs to install_flow.rs to improve code organization and separation of concerns. This refactoring does not change functionality but centralizes dependency installation logic.
